[FIX] TH181127/0009 product_brand_supplier: fix code error and PEP8

+++ product_brand_supplier/models/purchase.py
@@ -13,14 +13,26 @@ class ProcurementOrder(models.Model):
         cache = {}
         res = []
         for procurement in self:
+            new_supplierinfo = False
             suppliers = procurement.product_id.seller_ids\
-                .filtered(lambda r: (not r.company_id or r.company_id == procurement.company_id) and (not r.product_id or r.product_id == procurement.product_id))
+                .filtered(lambda r: (not r.company_id or
+                                     r.company_id == procurement.company_id)
+                          and (not r.product_id or r.product_id ==
+                               procurement.product_id))
             if not suppliers:
-                if procurement.product_id.product_brand_id.partner_id and procurement.product_id.product_brand_id.partner_id.supplier:
-                    new_supplierinfo = self.env['product.supplierinfo'].create({'product_tmpl_id': procurement.product_id.product_tmpl_id.id, 'name': procurement.product_id.product_brand_id.partner_id.id})
+                if procurement.product_id.product_brand_id.partner_id and \
+                        procurement.product_id.product_brand_id.partner_id.\
+                        supplier:
+                    new_supplierinfo = self.env['product.supplierinfo']\
+                        .create({'product_tmpl_id': procurement.product_id.
+                                 product_tmpl_id.id,
+                                 'name': procurement.product_id.
+                                 product_brand_id.partner_id.id})
                     suppliers = [new_supplierinfo]
                 else:
-                    procurement.message_post(body=_('No vendor associated to product %s. Please set one to fix this procurement.') % (procurement.product_id.name))
+                    procurement.message_post(body=_('No vendor associated to \
+                    product %s. Please set one to fix this procurement.')
+                                             % (procurement.product_id.name))
                     continue
             supplier = procurement._make_po_select_supplier(suppliers)
             partner = supplier.name
@@ -36,21 +48,30 @@ class ProcurementOrder(models.Model):
             if not po:
                 vals = procurement._prepare_purchase_order(partner)
                 po = self.env['purchase.order'].create(vals)
-                name = (procurement.group_id and (procurement.group_id.name + ":") or "") + (procurement.name != "/" and procurement.name or "")
-                message = _("This purchase order has been created from: <a href=# data-oe-model=procurement.order data-oe-id=%d>%s</a>") % (procurement.id, name)
+                name = (procurement.group_id and (procurement.group_id.name +
+                        ":") or "") + (procurement.name != "/" and
+                                       procurement.name or "")
+                message = _("This purchase order has been created from: \
+                    <a href=# data-oe-model=procurement.order \
+                    data-oe-id=%d>%s</a>") % (procurement.id, name)
                 po.message_post(body=message)
                 cache[domain] = po
-            elif not po.origin or procurement.origin not in po.origin.split(', '):
+            elif not po.origin or \
+                    procurement.origin not in po.origin.split(', '):
                 # Keep track of all procurements
                 if po.origin:
                     if procurement.origin:
-                        po.write({'origin': po.origin + ', ' + procurement.origin})
+                        po.write({'origin': po.origin + ', ' +
+                                  procurement.origin})
                     else:
                         po.write({'origin': po.origin})
                 else:
                     po.write({'origin': procurement.origin})
-                name = (self.group_id and (self.group_id.name + ":") or "") + (self.name != "/" and self.name or "")
-                message = _("This purchase order has been modified from: <a href=# data-oe-model=procurement.order data-oe-id=%d>%s</a>") % (procurement.id, name)
+                name = (self.group_id and (self.group_id.name + ":") or "") + \
+                    (self.name != "/" and self.name or "")
+                message = _("This purchase order has been modified from: \
+                    <a href=# data-oe-model=procurement.order \
+                    data-oe-id=%d>%s</a>") % (procurement.id, name)
                 po.message_post(body=message)
             if po:
                 res += [procurement.id]
@@ -58,20 +79,31 @@ class ProcurementOrder(models.Model):
             # Create Line
             po_line = False
             for line in po.order_line:
-                if line.product_id == procurement.product_id and line.product_uom == procurement.product_id.uom_po_id:
-                    procurement_uom_po_qty = procurement.product_uom._compute_quantity(procurement.product_qty, procurement.product_id.uom_po_id)
+                if line.product_id == procurement.product_id and \
+                        line.product_uom == procurement.product_id.uom_po_id:
+                    procurement_uom_po_qty = procurement.product_uom.\
+                        _compute_quantity(procurement.product_qty,
+                                          procurement.product_id.uom_po_id)
                     seller = procurement.product_id._select_seller(
                         partner_id=partner,
                         quantity=line.product_qty + procurement_uom_po_qty,
                         date=po.date_order and po.date_order[:10],
                         uom_id=procurement.product_id.uom_po_id)
 
-                    price_unit = self.env['account.tax']._fix_tax_included_price_company(seller.price, line.product_id.supplier_taxes_id, line.taxes_id, self.company_id) if seller else 0.0
-                    if price_unit and seller and po.currency_id and seller.currency_id != po.currency_id:
-                        price_unit = seller.currency_id.compute(price_unit, po.currency_id)
+                    price_unit = self.env['account.tax'].\
+                        _fix_tax_included_price_company(
+                            seller.price,
+                            line.product_id.supplier_taxes_id,
+                            line.taxes_id,
+                            self.company_id) if seller else 0.0
+                    if price_unit and seller and po.currency_id and \
+                            seller.currency_id != po.currency_id:
+                        price_unit = seller.currency_id.compute(price_unit,
+                                                                po.currency_id)
 
                     po_line = line.write({
-                        'product_qty': line.product_qty + procurement_uom_po_qty,
+                        'product_qty': line.product_qty +
+                        procurement_uom_po_qty,
                         'price_unit': price_unit,
                         'procurement_ids': [(4, procurement.id)]
                     })
@@ -79,5 +111,6 @@ class ProcurementOrder(models.Model):
             if not po_line:
                 vals = procurement._prepare_purchase_order_line(po, supplier)
                 self.env['purchase.order.line'].create(vals)
-            new_supplierinfo.unlink()
+            if new_supplierinfo:
+                new_supplierinfo.unlink()
         return res
